backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-multipart
python-dotenv>=0.19.0
httpx
sqlalchemy[asyncio]>=2.0.0
alembic
asyncpg
psycopg2-binary
pydantic>=2.0.0
pydantic-settings
starlette
anyio>=3.7.1,<4.0.0
google-generativeai==0.8.3
boto3>=1.26.0
botocore>=1.29.0
reportlab>=4.0.0
fpdf2>=2.7.0
